Minimize the .vimrc
The .vimrc has not been maintained since I made the move to neovim for
development. The config as-written prior to this commit caused vim to
immediately freeze when launched.

In an effort to ensure that vim continues to work on new machines, I have
decided to drastically reduce the .vimrc to include only a few of my basic
preferences (mostly around shortcuts and controls) so that it feels familiar
when I use it.

The plan is to overhaul the neovim init file soon to completely remove Vundle,
and the .vimrc will be kept around for small quality-of-life improvements in
situations when neovim is not available or needed for whatever reason.