Overhaul Neovim Configuration
After many years, I've finally been inspired by @ThePrimeagen and
@tjdevries' recent videos, and I'm actually updating my vim
configuration to be Neovim-specific.

I may resurrect my `.vimrc` if I find the need to duck into a more
tailored Vim experience, but I've been pretty heavily Neovim for a while
now.

Changes
- Replace [Vundle](https://github.com/VundleVim/Vundle.vim) with
  [packer.nvim](https://github.com/wbthomason/packer.nvim)
- Mostly get all plugins working
- Update `macos-setup.sh` to reflect my new repo & preferences

Not everything is quite perfect yet (PaperColor doesn't seem to be used
in all places), but this is good enough for a commit.