Add Makefile for Easier Sync (#3)
The main issue with the way this repo has been set up is that it is only
for initializing a development environment. Keeping it in-sync has been
more tedious.

To alleviate this, I'm going to try some Make commands. This will allow
me to update my Neovim config files in this repo, run `make
sync-neovim`, and then see those changes reflected. If I don't like the
result, I can revert to a previous commit and re-sync.