Add Initial `.zshrc` File (#4)
* Add Initial `.zshrc` File

I recently switched to iTerm2, which uses zsh. This is also the default
shell for MacOS' Terminal.

So, I've _started_ to port my (limited) bash configuration over to zsh.

This has conda stuff in it, and I'm not sure if I added that or if it's
from some other package that I added... So, I'm leaving it for now. My
Python installation needs a mature setup and overhaul anyway.

* Add `sync-zsh` to `sync` command

Since I removed the broken `source` commands, I can just run `make sync`
anytime to ensure that everything is aligned.