Fix Setup w/ Bash & Brew for M1
I've split out the setup and maintenance commands because the setup
script needs to be run in a Bash terminal, but I want the maintenance
script (for updating packages & syncing configuration) to be
shell-agnostic.

There were also some issues with running Homebrew and treesitter on an
M1 chip, so those required fresh installs + a magic shell `eval`
command, which I've added.