Fix Python Installation & Other Stuff
Frankly, I've never bothered to learn how Python installations work, and
I don't really care to. I just want to be able to run `python` in the
terminal, and I was briefly unable to do so - as only `python3` was
working.

This fixes that issue.

Notable change is updating the `PATH` for Applie Silicon, I believe.